Separate table_test.go from sqlfile_test.go
